docker/compile-neuron/requirements-test.txt
boto3==1.26.77
torchtext==0.14.1
transformers>=4.25.1 # or <= 4.20. See for details: https://github.com/aws-neuron/aws-neuron-sdk/issues/593
pytest==7.2.1